+++ lshell/variables.py
@@ -3,7 +3,7 @@
 import sys
 import os
 
-__version__ = "0.11.5"
+__version__ = "0.11.6"
 
 # Required config variable list per user
 required_config = ["allowed", "forbidden", "warning_counter"]